Inzynieria oprogramowania/project/front/requirements.txt
tkinter==8.6
pyodbc==4.0.30
Pillow==8.0.1